update code for image description
